+++ example/lib/pages/one_video_ctrl.dart
@@ -1,23 +1,49 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:video_box/video.controller.dart';
 import 'package:video_box/video_box.dart';
 import 'package:video_player/video_player.dart';
 
-import 'globals.dart';
+import '../globals.dart';
+
+class MyFullScreen implements CustomFullScreen {
+  @override
+  void close(BuildContext context, VideoController controller) {
+    Navigator.of(context).pop(controller.value.positionText);
+  }
+
+  @override
+  Future open(BuildContext context, VideoController controller) async {
+    setLandscape();
+    SystemChrome.setEnabledSystemUIOverlays([]);
+    await Navigator.of(context).push<String>(
+      MaterialPageRoute(
+        builder: (_) {
+          return Scaffold(
+            body: Center(child: VideoBox(controller: controller)),
+          );
+        },
+      ),
+    ).then(print);
+    setPortrait();
+    SystemChrome.setEnabledSystemUIOverlays(SystemUiOverlay.values);
+  }
+}
 
 class OneVideoCtrl extends StatefulWidget {
   @override
   _OneVideoCtrlState createState() => _OneVideoCtrlState();
 }
 
-class _OneVideoCtrlState extends State<OneVideoCtrl> {
+class _OneVideoCtrlState extends State<OneVideoCtrl>
+    {
   VideoController vc;
-  bool get isPlay => vc.videoCtrl.value.isPlaying;
   ScrollController controller = ScrollController();
 
   @override
   void initState() {
     super.initState();
+    
     vc = VideoController(
       source: VideoPlayerController.network(src1),
       loop: true,
@@ -28,11 +54,43 @@ class _OneVideoCtrlState extends State<OneVideoCtrl> {
       inactiveColor: Colors.pink,
       background: Colors.indigo,
       circularProgressIndicatorColor: Colors.lime,
+      bottomPadding: EdgeInsets.only(bottom: 10),
+      customLoadingWidget: Center(
+        child: Container(
+          color: Colors.pink,
+          padding: const EdgeInsets.all(12),
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: <Widget>[
+              CircularProgressIndicator(),
+              SizedBox(height: 12),
+              Text("Lading..."),
+            ],
+          ),
+        ),
+      ),
+      customBufferedWidget: Center(
+        child: Container(
+          color: Colors.pink,
+          padding: const EdgeInsets.all(12),
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: <Widget>[
+              CircularProgressIndicator(),
+              SizedBox(height: 12),
+              Text("please wait..."),
+            ],
+          ),
+        ),
+      ),
+      customFullScreen: MyFullScreen(),
       // cover: Image.network('https://i.loli.net/2019/08/29/7eXVLcHAhtO9YQg.jpg'),
       // controllerWidgets: false,
       // cover: Text('Cover'),
       // initPosition: Duration(minutes: 23, seconds: 50)
-    )..initialize().then((_) {
+    )
+      ..addFullScreenChangeListener((c) async {})
+      ..initialize().then((_) {
         // initialized
       });
   }
@@ -42,6 +100,7 @@ class _OneVideoCtrlState extends State<OneVideoCtrl> {
     vc.dispose();
     super.dispose();
   }
+
 
   @override
   Widget build(BuildContext context) {
@@ -87,20 +146,7 @@ class _OneVideoCtrlState extends State<OneVideoCtrl> {
               RaisedButton(
                 child: Text('full screen'),
                 onPressed: () {
-                  vc.onFullScreen(
-                    context,
-                    SafeArea(
-                      child: Scaffold(
-                        body: Center(
-                          child: AspectRatio(
-                            aspectRatio: 16 / 9,
-                            child: VideoBox(controller: vc),
-                          ),
-                        ),
-                        backgroundColor: Colors.black,
-                      ),
-                    ),
-                  );
+                  vc.onFullScreenSwitch(context);
                 },
               ),
               RaisedButton(