+++ lib/video.controller.dart
@@ -3,13 +3,45 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+
 import 'package:mobx/mobx.dart';
 import 'package:video_player/video_player.dart';
-import 'package:video_box/util/duration_string.dart' show durationString;
 import 'package:volume/volume.dart';
 
-import 'video_box.dart';
+import 'video_box.dart' show CustomFullScreen, VideoBoxFullScreenPage;
+import 'util/duration_string.dart' show durationString;
+import 'video_state.dart';
+
 part 'video.controller.g.dart';
+
+typedef FullScreenChange = Function(VideoController controller);
+
+Route<T> _defaultCustomFullScreenRoute<T>(VideoController controller) {
+  return MaterialPageRoute<T>(
+      builder: (_) => VideoBoxFullScreenPage(controller: controller));
+}
+
+/// 设置为正常模式
+///
+/// Set to normal mode
+void setPortrait() {
+  SystemChrome.setPreferredOrientations([
+    DeviceOrientation.landscapeRight,
+    DeviceOrientation.landscapeLeft,
+    DeviceOrientation.portraitUp,
+    DeviceOrientation.portraitDown,
+  ]);
+}
+
+/// 设置为横屏模式
+///
+/// Set to landscape mode
+void setLandscape() {
+  SystemChrome.setPreferredOrientations([
+    DeviceOrientation.landscapeRight,
+    DeviceOrientation.landscapeLeft,
+  ]);
+}
 
 class VideoController = _VideoController with _$VideoController;
 
@@ -32,6 +64,10 @@ abstract class _VideoController with Store {
     this.circularProgressIndicatorColor = Colors.white,
     this.animetedIconDuration = kTabScrollDuration,
     Color barrierColor,
+    this.customLoadingWidget,
+    this.customBufferedWidget,
+    this.customFullScreen,
+    this.bottomPadding = EdgeInsets.zero,
   }) {
     videoCtrl = source;
     this.barrierColor = barrierColor ?? Colors.black.withOpacity(0.6);
@@ -45,32 +81,121 @@ abstract class _VideoController with Store {
   final Color inactiveColor;
   final Color circularProgressIndicatorColor;
 
+  /// 底部控制器的padding
+  /// 
+  /// Padding of bottom controller
+  final EdgeInsets bottomPadding;
+
+  ///
+  ///Example
+  ///```dart
+  ///class MyFullScreen extends CustomFullScreen {
+  ///   @override
+  ///   void close(BuildContext context, VideoController controller) {
+  ///     Navigator.of(context).pop(controller.value.positionText);
+  ///   }
+  ///
+  ///   @override
+  ///   Future open(BuildContext context, VideoController controller) async {
+  ///     setLandscape();
+  ///     SystemChrome.setEnabledSystemUIOverlays([]);
+  ///     await Navigator.of(context)
+  ///         .push<String>(
+  ///           PageRouteBuilder(
+  ///             transitionDuration: Duration(seconds: 2),
+  ///             pageBuilder: (context, animation, secondaryAnimation) {
+  ///               return Scaffold(
+  ///                   body: Center(child: VideoBox(controller: controller)));
+  ///             },
+  ///             transitionsBuilder:
+  ///                 (context, animation, secondaryAnimation, child) {
+  ///               return ScaleTransition(
+  ///                 scale: Tween<double>(begin: 0.0, end: 1.0).animate(
+  ///                   CurvedAnimation(
+  ///                     parent: animation,
+  ///                     curve: Curves.ease,
+  ///                   ),
+  ///                 ),
+  ///                 child: child,
+  ///               );
+  ///             },
+  ///           ),
+  ///         )
+  ///         .then(print);
+  ///     setPortrait();
+  ///     SystemChrome.setEnabledSystemUIOverlays(SystemUiOverlay.values);
+  ///   }
+  /// }
+  ///```
+  final CustomFullScreen customFullScreen;
+
+  /// 当视频首次加载时将显示这个widget。
+  ///
+  /// This widget will be displayed when the video is first loaded.
+  ///
+  /// Example:
+  /// ```dart
+  /// vc = VideoController(
+  ///   ...
+  ///   customLoadingWidget: Center(
+  ///     child: Container(
+  ///       color: Colors.pink,
+  ///       padding: const EdgeInsets.all(12),
+  ///       child: Column(
+  ///         mainAxisSize: MainAxisSize.min,
+  ///         children: <Widget>[
+  ///           CircularProgressIndicator(),
+  ///           SizedBox(height: 12),
+  ///           Text("Lading..."),
+  ///         ],
+  ///       ),
+  ///     ),
+  ///   ),
+  /// )
+  /// ```
+  final Widget customLoadingWidget;
+
+  /// 当视频进入缓冲时将显示这个widget。
+  /// 类似[customLoadingWidget]那样设置
+  ///
+  /// This widget will be displayed when the video enters the buffer.
+  /// Set like [customLoadingWidget]
+  final Widget customBufferedWidget;
+
   /// 监听视频播放回调
+  ///
   /// Listening to video playback callbacks
   Function _playingListenner;
   addListener(void Function() listener) {
-    this._playingListenner = listener;
+    _playingListenner = listener;
   }
 
   /// 监听视频播放结束回调
+  ///
   /// Listening to video playback end callback
   Function _playEnd;
   addPlayEndListener(void Function() listener) {
-    this._playEnd = listener;
+    _playEnd = listener;
   }
 
-  Function _fullScreenChange;
-  addFullScreenChangeListener(void Function() listener) {
-    this._fullScreenChange = listener;
+  FullScreenChange _fullScreenChange;
+
+  /// 添加监听屏幕改变事件
+  ///
+  /// Add listen screen change events
+  addFullScreenChangeListener(FullScreenChange listener) {
+    _fullScreenChange = listener;
   }
 
   /// 控制媒体音量
+  ///
   /// Controlling media volume
   Future<void> _initPlatformState() async {
     await Volume.controlVolume(AudioManager.STREAM_MUSIC);
   }
 
   /// 动画icon的初始化
+  ///
   /// Initialization of the animation icon
   void initAnimetedIconController(TickerProvider vsync) {
     animetedIconController = AnimationController(
@@ -102,6 +227,7 @@ abstract class _VideoController with Store {
   List<Widget> afterChildren;
 
   /// 控制器背景色
+  ///
   /// Controller background color
   @observable
   Color barrierColor;
@@ -109,6 +235,7 @@ abstract class _VideoController with Store {
   bool isPlayEnd = false;
 
   /// 自动关闭 Controller layer 的计时器
+  ///
   /// Automatically turn off the timer of the Controller layer
   Timer _controllerLayerTimer;
 
@@ -120,6 +247,7 @@ abstract class _VideoController with Store {
   Widget cover;
 
   /// 是否显示默认控件 默认[true]
+  ///
   /// Whether to show default controls default [true]
   @observable
   bool controllerWidgets;
@@ -134,6 +262,7 @@ abstract class _VideoController with Store {
   }
 
   /// 随时监听缓冲状态
+  ///
   /// Listen to buffer status at any time
   @action
   void setVideoBuffer() {
@@ -158,6 +287,7 @@ abstract class _VideoController with Store {
   void setCover(Widget newCover) => cover = newCover;
 
   /// 是否显示封面，只有在第一次播放前显示
+  ///
   /// Whether to show the cover, only before the first play
   @computed
   bool get isShowCover {
@@ -210,11 +340,13 @@ abstract class _VideoController with Store {
   Duration position;
 
   /// 视频总时长
+  ///
   /// Total video duration
   @observable
   Duration duration;
 
   /// 是否显示控制器层
+  ///
   /// Whether to show the controller layer
   @observable
   bool controllerLayer = true;
@@ -241,11 +373,19 @@ abstract class _VideoController with Store {
   void toggleShowVideoCtrl() => setControllerLayer(show: !controllerLayer);
 
   /// 是否为全屏播放
+  ///
   /// Whether to play in full screen
   @observable
   bool isFullScreen = false;
 
+  @action
+  void _setFullScreen({bool full}) {
+    isFullScreen = full;
+    if (_fullScreenChange != null) _fullScreenChange(this);
+  }
+
   /// 快进，快退的时间
+  ///
   /// Fast forward and rewind time
   @observable
   Duration skiptime;
@@ -257,11 +397,13 @@ abstract class _VideoController with Store {
   }
 
   /// 25:00 or 2:00:00 总时长
+  ///
   /// Total duration
   @computed
   String get durationText => duration == null ? '' : durationString(duration);
 
   /// 00:01 当前时间
+  ///
   /// current time
   @computed
   String get positionText =>
@@ -282,6 +424,7 @@ abstract class _VideoController with Store {
   double sliderBufferValue = 0.0;
 
   /// 返回一个符合当前音量的icon
+  ///
   /// Returns an icon that matches the current volume
   @computed
   IconData get volumeIcon => volume <= 0
@@ -293,6 +436,7 @@ abstract class _VideoController with Store {
       !isFullScreen ? Icons.fullscreen : Icons.fullscreen_exit;
 
   /// 替换当前播放的视频资源
+  ///
   /// Replace the currently playing video resource
   @action
   Future<void> setSource(VideoPlayerController source) async {
@@ -304,10 +448,12 @@ abstract class _VideoController with Store {
   }
 
   /// 控制器层存活的时间
+  ///
   /// Controller layer time to live
   final Duration controllerDuration;
 
   /// 初始化viedo控制器
+  ///
   /// Initialize the viedo controller
   @action
   Future<void> initialize() async {
@@ -332,6 +478,7 @@ abstract class _VideoController with Store {
   }
 
   /// 视频播放时的监听器
+  ///
   /// Listener during video playback
   ///
   /// seek 也会触发
@@ -365,6 +512,7 @@ abstract class _VideoController with Store {
   }
 
   /// 开启声音或关闭
+  ///
   /// Turn sound on or off
   void setOnSoundOrOff() {
     if (videoCtrl.value != null) {
@@ -374,6 +522,7 @@ abstract class _VideoController with Store {
   }
 
   /// 播放或暂停
+  ///
   /// Play or pause
   Future<void> togglePlay() async {
     // 等待Icon动画关闭
@@ -413,6 +562,7 @@ abstract class _VideoController with Store {
   }
 
   /// 控制播放时间位置
+  ///
   /// Controlling playback time position
   Future<void> seekTo(Duration d) async {
     if (videoCtrl.value != null) {
@@ -434,58 +584,38 @@ abstract class _VideoController with Store {
   }
 
   /// 进度条被改变
+  ///
   /// The progress bar is changed
   void sliderChanged(double v) =>
       seekTo(Duration(seconds: (v * duration.inSeconds).toInt()));
 
-  /// 设置为横屏模式
-  /// Set to landscape mode
-  @action
-  void _setLandscape() {
-    isFullScreen = true;
-    SystemChrome.setPreferredOrientations([
-      DeviceOrientation.landscapeRight,
-      DeviceOrientation.landscapeLeft,
-    ]);
-  }
-
-  /// 设置为正常模式
-  /// Set to normal mode
-  @action
-  void _setPortrait() {
-    isFullScreen = false;
-    SystemChrome.setPreferredOrientations([
-      DeviceOrientation.landscapeRight,
-      DeviceOrientation.landscapeLeft,
-      DeviceOrientation.portraitUp,
-      DeviceOrientation.portraitDown,
-    ]);
-  }
-
-  /// screen  自定义全屏page
-  Future<void> onFullScreen(BuildContext context, [Widget customScreen]) async {
-    if (isFullScreen) {
-      // Exit Full Screen
-      Navigator.of(context).pop();
+  /// 打开或关闭全屏
+  ///
+  /// Turn full screen on or off
+  Future<void> onFullScreenSwitch(BuildContext context) async {
+    if (customFullScreen == null) {
+      if (isFullScreen) {
+        Navigator.of(context).pop();
+      } else {
+        _setFullScreen(full: true);
+        await Navigator.of(context).push(_defaultCustomFullScreenRoute(this));
+        // 这里可以监听到系统导航栏的返回事件
+        _setFullScreen(full: false);
+        setPortrait();
+      }
     } else {
-      // Turn on full screen
-      _setLandscape();
-      // Screen.keepOn(true);
-      SystemChrome.setEnabledSystemUIOverlays([]);
-      if (_fullScreenChange != null) _fullScreenChange();
-      await Navigator.of(context).push(
-        MaterialPageRoute(
-          builder: (_) => customScreen ?? _FullPageVideo(controller: this),
-        ),
-      );
-      _setPortrait();
-      // Screen.keepOn(false);
-      SystemChrome.setEnabledSystemUIOverlays(SystemUiOverlay.values);
-      if (_fullScreenChange != null) _fullScreenChange();
+      if (isFullScreen) {
+        customFullScreen.close(context, this);
+      } else {
+        _setFullScreen(full: true);
+        await customFullScreen.open(context, this);
+        _setFullScreen(full: false);
+      }
     }
   }
 
   /// 右侧版块设置媒体音量
+  ///
   /// Set the media volume on the right side
   Future<void> setMediaVolume(DragUpdateDetails d) async {
     // if (d.delta.dy < 0) {
@@ -501,8 +631,10 @@ abstract class _VideoController with Store {
   // double _currentBrightness;
 
   /// 左侧版块设置屏幕亮度
+  ///
   /// The left section sets the screen brightness
-  /// TODO: Screen 出现错误删除此依赖，寻找其它解决方案
+  /// ! [screen 0.0.5]出现错误，删除此依赖，寻找其它解决方案
+  /// https://pub.dev/packages/screen
   Future<void> setScreenBrightness(DragUpdateDetails d) async {
     // double dy = d.delta.dy / 200;
     // _currentBrightness ??= await Screen.brightness;
@@ -519,7 +651,6 @@ abstract class _VideoController with Store {
   /// set MEDIA Volume
   Future<int> setVol(int i) => Volume.setVol(i);
 
-  /// 清理资源
   Future<void> dispose() async {
     animetedIconController?.dispose();
     videoCtrl?.removeListener(_videoListenner);
@@ -547,89 +678,4 @@ abstract class _VideoController with Store {
 
   @override
   String toString() => value.toString();
-}
-
-class VideoState {
-  VideoState({
-    this.dataSource,
-    this.dataSourceType,
-    this.size,
-    this.isLooping,
-    this.isPlaying,
-    this.autoplay,
-    this.volume,
-    this.initPosition,
-    this.position,
-    this.duration,
-    this.skiptime,
-    this.positionText,
-    this.durationText,
-    this.sliderValue,
-  });
-
-  String dataSource;
-  DataSourceType dataSourceType;
-  Size size;
-  bool autoplay;
-  bool isLooping;
-  bool isPlaying;
-  double volume;
-  Duration initPosition;
-  Duration position;
-  Duration duration;
-  Duration skiptime;
-  String positionText;
-  String durationText;
-  double sliderValue;
-
-  Map<String, dynamic> toMap() {
-    return {
-      "dataSource": dataSource,
-      "dataSourceType": dataSourceType,
-      "size": size,
-      "autoplay": autoplay,
-      "isLooping": isLooping,
-      "isPlaying": isPlaying,
-      "volume": volume,
-      "initPosition": initPosition,
-      "position": position,
-      "duration": duration,
-      "skiptime": skiptime,
-      "positionText": positionText,
-      "durationText": durationText,
-      "sliderValue": sliderValue,
-    };
-  }
-
-  @override
-  String toString() {
-    return """
-    {
-      "dataSource": $dataSource,
-      "dataSourceType": $dataSourceType,
-      "size": $size,
-      "autoplay": $autoplay,
-      "isLooping": $isLooping,
-      "isPlaying": $isPlaying,
-      "volume": $volume,
-      "initPosition": $initPosition,
-      "position": $position,
-      "duration": $duration,
-      "skiptime": $skiptime,
-      "positionText": $positionText,
-      "durationText": $durationText,
-      "sliderValue": $sliderValue,
-    }
-    """;
-  }
-}
-
-class _FullPageVideo extends StatelessWidget {
-  final VideoController controller;
-
-  const _FullPageVideo({Key key, this.controller}) : super(key: key);
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(body: Center(child: VideoBox(controller: controller)));
-  }
 }