+++ lib/widgets/video_bottom_ctroller.dart
@@ -31,7 +31,7 @@ class VideoBottomCtroller extends StatelessWidget {
           Observer(
             builder: (_) => IconButton(
               icon: Icon(controller.fullScreenIcon),
-              onPressed: () => controller.onFullScreen(context),
+              onPressed: () => controller.onFullScreenSwitch(context),
             ),
           ),
         ],